pyproject.toml
[project]
name = "working-hands"
version = "0.1.0"
description = "Test task"
authors = [
    {name = "Daniil Neudakhin",email = "neudakhin.perm@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.9,<3.10"
dependencies = [
    "django (==4.2.17)",
    "django-rest-framework (>=0.1.0,<0.2.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "django-extensions (>=4.1,<5.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
